Renamed python source to hungarian_match and recursive_match to be consistent with rust implementation. Documentation changes to include changelog. Sphinx documentation changes.
