+++ todo.txt
@@ -14,3 +14,6 @@
     - CMake
 
 - Paper and analysis on the recursive algorithm. 
+
+v1.3 - C implementation
+v2.0 - Python class implementation and cleanups
